Removed Refresh button for now. Moved ViewUtility functions to View class and Adapter class. Removed No Results found for now. Cleaned up code
