+++ app/build.gradle
@@ -49,6 +49,8 @@ dependencies {
     testImplementation 'io.mockk:mockk:1.10.2'
 
     testImplementation "androidx.arch.core:core-testing:2.1.0"
+
+    // Glide
     implementation 'com.github.bumptech.glide:glide:4.14.2'
     kapt 'com.github.bumptech.glide:compiler:4.11.0'
