+++ app/src/main/java/com/samm/practiceapp01/core/Constants.kt
@@ -1,4 +1,4 @@
-package com.samm.practiceapp01.util
+package com.samm.practiceapp01.core
 
 object Constants {
     const val API_KEY = "9ea6d6be1e6d41d2a86042774ef87848"
