+++ app/src/main/java/com/samm/practiceapp01/data/RepositoryImpl.kt
@@ -5,15 +5,15 @@ import androidx.lifecycle.MutableLiveData
 import com.samm.practiceapp01.data.database.NewsDao
 import com.samm.practiceapp01.domain.Repository
 import com.samm.practiceapp01.domain.models.NewsItem
-import com.samm.practiceapp01.util.Constants.PAGE_SIZE
+import com.samm.practiceapp01.core.Constants.PAGE_SIZE
 import org.json.JSONObject
 import retrofit2.Response
 
 class RepositoryImpl (private val api: NewsApi, private val myDao: NewsDao): Repository {
 
+    private var errorString: String = ""
     val loading: MutableLiveData<Boolean> = MutableLiveData()
     val errorMessageLD: MutableLiveData<String> = MutableLiveData()
-    val noResults: MutableLiveData<Boolean> = MutableLiveData()
     val getNewsFromDatabase = myDao.getAllNewsItems()
 
     override suspend fun clearCache() {
@@ -21,32 +21,34 @@ class RepositoryImpl (private val api: NewsApi, private val myDao: NewsDao): Rep
     }
     override suspend fun fetchArticles(search: String, page: Int) {
         loading.postValue(true)
-
         try {
             val response = getNews(search, page)
-            Log.d("Response?", response.isSuccessful.toString())
-            if (response.isSuccessful) {
-                noResults.postValue(false)
-                response.body()?.let { newsItem ->
-                    clearCache()
-                    addArticleToDatabase(newsItem)
+            when {
+                response.isSuccessful -> {
+                    response.body()?.let { newsItem ->
+                        clearCache()
+                        addArticleToDatabase(newsItem)
+                    }
                 }
-
-            } else {
-                val parsedErrorResponse = parseErrorBody(response)
-                errorMessageLD.postValue(parsedErrorResponse)
-                noResults.postValue(true)
-                Log.d("Error Response", parsedErrorResponse)
+                response.errorBody() != null -> {
+                    val parsedErrorResponse = parseErrorBody(response)
+                    errorString = parsedErrorResponse
+                    errorMessageLD.postValue(errorString)
+                }
             }
         } catch (e: Exception){
+            errorMessageLD.postValue(errorString)
             Log.d("Exception:", "$e")
         }
         loading.postValue(false)
     }
 
-    // API call
     override suspend fun getNews(search: String, page: Int): Response<NewsItem> {
-        return api.getNews(searchQuery = search, pageNumber = page, pageSize = PAGE_SIZE)
+        return api.getNews(
+            searchQuery = search,
+            pageNumber = page,
+            pageSize = PAGE_SIZE
+        )
     }
 
     override suspend fun addArticleToDatabase(newsItem: NewsItem){
@@ -55,7 +57,6 @@ class RepositoryImpl (private val api: NewsApi, private val myDao: NewsDao): Rep
         }
     }
 
-    // parse the error body from the API response
     override fun parseErrorBody(response: Response<NewsItem>): String {
         val errorJson = JSONObject(response.errorBody()?.string() ?: "")
         val errorMessage = errorJson.getString("message")
