+++ app/src/main/java/com/samm/practiceapp01/presentation/ArticleFragment.kt
@@ -6,19 +6,21 @@ import android.view.*
 import android.widget.*
 import androidx.core.view.MenuHost
 import androidx.core.view.MenuProvider
+import androidx.core.view.isEmpty
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.LayoutManager
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.samm.practiceapp01.R
-import com.samm.practiceapp01.util.ViewUtility
+import com.samm.practiceapp01.core.ViewUtility
 
 class ArticleFragment : Fragment() {
 
-    private val viewUtility = ViewUtility()
+    val viewUtility = ViewUtility()
     private lateinit var newsViewModel: NewsViewModel
     private lateinit var adapter: NewsAdapter
     private lateinit var recyclerView: RecyclerView
@@ -28,8 +30,7 @@ class ArticleFragment : Fragment() {
     private lateinit var layoutManager: LayoutManager
     private lateinit var errorMessageTV: TextView
     private var pageNumber = 1
-    private var results = false
-    private var observeResults: Boolean = false
+
     @SuppressLint("MissingInflatedId")
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -55,7 +56,6 @@ class ArticleFragment : Fragment() {
         newsViewModel = ViewModelProvider(this)[NewsViewModel::class.java]
 
 
-
         return view
     }
 
@@ -64,20 +64,16 @@ class ArticleFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         progressBar.visibility = View.GONE
         setUpRecyclerView(layoutManager)
-        ifErrorMessage()
-        observeResults = observeResults(newsViewModel)
-
         newsViewModel.newsData(viewLifecycleOwner, adapter)
-        viewUtility.hideViewsWhenScrolled(recyclerView, searchField, backToTopButton)
-
+        hideViewsWhenScrolled(recyclerView, searchField, backToTopButton)
         backToTopButton.setOnClickListener { backToTopButtonClickListener() }
-
         // Get News Data when search button in keyboard is clicked
         searchField.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
             override fun onQueryTextSubmit(query: String?): Boolean {
                 if (query != null) {
                     pageNumber = 1
                     fetchNewsData(pageNumber, query)
+                    ifErrorMessage()
                 }
                 return false
             }
@@ -87,7 +83,6 @@ class ArticleFragment : Fragment() {
                 return false
             }
         })
-
         val menuHost: MenuHost = requireActivity()
         menuHost.addMenuProvider(object : MenuProvider {
             override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
@@ -95,16 +90,13 @@ class ArticleFragment : Fragment() {
             }
 
             override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
-                // Handle item selection
                 return when (menuItem.itemId) {
-                    R.id.refresh -> {
-                        viewUtility.observeLoading(newsViewModel, viewLifecycleOwner, progressBar)
-                        newsViewModel.getArticles(pageNumber, searchField.query.toString())
-                        true
-                    }
                     R.id.clear -> {
-                        newsViewModel.clearCache()
-                        adapter.clearList()
+                        newsViewModel.deleteAllAlertDialog(requireContext()){
+                            newsViewModel.clearCache()
+                            adapter.clearList()
+                        }
+
                         true
                     }
                     else -> false
@@ -124,28 +116,18 @@ class ArticleFragment : Fragment() {
 
     private fun fetchNewsData(page: Int, search: String) {
 
-        if(search.isEmpty()){
+        if (search.isEmpty()) {
             Toast.makeText(
                 activity, "Please enter a search term", Toast.LENGTH_LONG
             ).show()
         } else {
             // Get Data
-            viewUtility.observeLoading(newsViewModel, viewLifecycleOwner, progressBar)
+            observeLoading(newsViewModel, viewLifecycleOwner, progressBar, recyclerView)
             newsViewModel.getArticles(page, search)
-            observeResults(newsViewModel)
             viewUtility.hideKeyboard(activity)
         }
     }
 
-    private fun observeResults(viewModel: NewsViewModel): Boolean {
-        viewModel.noResults.observe(viewLifecycleOwner){ noResults ->
-            results = noResults
-        }
-        if (results){
-            Toast.makeText(context, "No Results", Toast.LENGTH_LONG).show()
-        }
-        return results
-    }
 
     private fun backToTopButtonClickListener() {
         recyclerView.smoothScrollToPosition(0)
@@ -155,8 +137,51 @@ class ArticleFragment : Fragment() {
     // Don't know if this is really working
     private fun ifErrorMessage(){
         newsViewModel.error.observe(viewLifecycleOwner){ error ->
-            errorMessageTV.text = error
             errorMessageTV.visibility = View.VISIBLE
+            errorMessageTV.text = error
+        }
+    }
+
+    private fun hideViewsWhenScrolled(
+        recyclerView: RecyclerView,
+        toolbar: View,
+        backToTopButton: FloatingActionButton
+    ){
+        recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
+                super.onScrolled(recyclerView, dx, dy)
+
+                // Get the first visible item position
+                val firstVisibleItemPosition = (recyclerView.layoutManager as LinearLayoutManager)
+                    .findFirstVisibleItemPosition()
+
+                // Check if the first visible item is the first item in the list
+                if (firstVisibleItemPosition == 0 || recyclerView.isEmpty()) {
+                    toolbar.visibility = View.VISIBLE
+                    backToTopButton.visibility = View.GONE
+                } else {
+                    toolbar.visibility = View.GONE
+                    backToTopButton.visibility = View.VISIBLE
+                }
+            }
+        })
+    }
+
+    private fun observeLoading(
+        newsViewModel: NewsViewModel,
+        viewLifecycleOwner: LifecycleOwner,
+        progressBar: ProgressBar,
+        recyclerView: RecyclerView
+    ){
+        newsViewModel.loading.observe(viewLifecycleOwner) { isLoading ->
+            if (isLoading) {
+                progressBar.visibility = View.VISIBLE
+                recyclerView.visibility = View.GONE
+
+            } else {
+                progressBar.visibility = View.GONE
+                recyclerView.visibility = View.VISIBLE
+            }
         }
     }
 }
