+++ app/src/main/java/com/samm/practiceapp01/presentation/NewsViewModel.kt
@@ -1,6 +1,8 @@
 package com.samm.practiceapp01.presentation
 
 import android.app.Application
+import android.content.Context
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
@@ -18,7 +20,6 @@ class NewsViewModel(application: Application) : AndroidViewModel(application) {
 
     val loading: MutableLiveData<Boolean> = repositoryImpl.loading
     val error: MutableLiveData<String> = repositoryImpl.errorMessageLD
-    val noResults: MutableLiveData<Boolean> = repositoryImpl.noResults
 
     private fun fetchArticles(search: String, page: Int) = viewModelScope.launch(Dispatchers.IO)  {
         repositoryImpl.fetchArticles(search, page)
@@ -37,10 +38,26 @@ class NewsViewModel(application: Application) : AndroidViewModel(application) {
     fun newsData(owner: LifecycleOwner, adapter: NewsAdapter){
         val newsList = repositoryImpl.getNewsFromDatabase
         newsList.observe(owner){ list ->
-            when {
-                list.isEmpty() -> noResults.postValue(true)
-                else -> adapter.setNews(list)
-            }
+            adapter.setNews(list)
         }
+    }
+
+    fun deleteAllAlertDialog(
+        context: Context,
+        deleteAll: () -> Unit
+    ){
+        val dialogBuilder = AlertDialog.Builder(context)
+        dialogBuilder.setMessage("Are you sure you want to clear the list?")
+            .setCancelable(false)
+            .setPositiveButton("Clear") { dialog, _ ->
+                deleteAll()
+                dialog.dismiss()
+            }
+            .setNegativeButton("Cancel") { dialog, _ ->
+                dialog.cancel()
+            }
+        val alert = dialogBuilder.create()
+        alert.setTitle("Warning")
+        alert.show()
     }
 }
