+++ app/src/main/res/menu/action_bar_menu.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/refresh"
-        android:title="Refresh" />
-    <item
         android:id="@+id/clear"
-        android:title="Clear" />
+        android:title="@string/clear" />
 </menu>
